dataset_clean functions
Created the outstanding functions needed to clean the dataset. Some small tweaks to dataset_clean.R to facilitate the changes. QA is filtering more than it should at the moment so will need some further tweaking but functions all run in the way we expact so have something to work with moving forward.